GCC 2.7.2.3 compatibility.
[git-p4: depot-paths = "//depot/qt/main/": change = 22208]

+++ src/widgets/qworkspace.cpp
@@ -2422,7 +2422,7 @@ void QWorkspaceChildTitleLabel::drawLabel()
 	    rS+=rD;
 	    gS+=gD;
 	    bS+=bD;
-	    p.setPen( QColor( rS, gS, bS ) );
+	    p.setPen( QColor( (int)rS, (int)gS, (int)bS ) );
 	    p.drawLine( x, contentsRect().y(), x, contentsRect().height() );
 	}
     } else {
